Added ASE and GPAW Installation Tests
To test the installation of ASE and GPAW on the cluster using a virtual
environment, it is helpful if the user has some very simple scripts to
run that utilize expected functions. The first folder provideis that as
a test that creates a water layer and adds that water layer to a nickel
cell. The second folder utilizes GPAW to optimize the N-N distance and
save a figure of the energy vs distance. Both taken from the ASE
tutorials at
https://wiki.fysik.dtu.dk/ase/gettingstarted/gettingstarted.html.

Co-authored-by: Rebecca Gerber <r-e-g-ular@users.noreply.github.com>