Define enviroments as extras; pin all deps.
Defining the various environments (e.g., `dev`, `test`, `docs`) in this
way allows developers to easily setup their development environment with
pip instead of requiring that Hatch be installed.

The main dependencies of the welcome guide are catered to a beginner user.
For this reason, the development dependencies are separate. When a user
runs `pip install .`, they will install only what is required to complete
the tutorials. The development dependencies can be installed as extras.

The use of ruff removes the need for black as a formatter.

Pinning all dependencies enforces compatible package versions with pip.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>