Fix software scripts; add where-to-edit comments
Fix sed command: the MacOS implementation of sed (BSD sed) requires an argument
for the -i option denoting the extension for backup files, but the version
installed on DRA clusters (GNU sed) does not. This causes sed to interpret
the sed script as a file on DRA clusters and causes an error.

The new comments direct users to where they should edit the software scripts
in the case that they download files individually.

Commands that append lines to files are edited to properly add newlines.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>